feat(): greatly simplified service workers
