+++ apps/pwabuilder-vscode/src/services/validation/sw-view.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import * as path from "path";
 import { readFile } from "fs/promises";
-import { findWorker } from "../service-worker";
+import { findWorker } from "../simple-service-worker";
 import { pathExists } from "../../library/file-utils";
 
 /**
